provide meter filter as bean
