+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/config/TkmsAutoConfiguration.java
@@ -37,6 +37,8 @@ public class TkmsAutoConfiguration {
     return new MeterCache(meterRegistry);
   }
 
+
+
   @Bean
   @ConditionalOnMissingBean(ITransactionsHelper.class)
   public TransactionsHelper twTransactionsHelper() {
